+++ cnccode/business_classes/BUCustomerSrAnalysisReport.inc.php
@@ -44,13 +44,13 @@ class BUCustomerSrAnalysisReport extends Business
         return $this->getResultsByDateRange($customerID, $fromDate, $toDate);
     }
 
-    function getResultsByPeriodRange($customerID, $startPeriod, $endPeriod)
+    function getResultsByPeriodRange($customerID, DateTimeInterface $startPeriod, DateTimeInterface $endPeriod)
     {
         /*
         Turn periods into dates
         */
-        $fromDate = $startPeriod . '-01';
-        $toDate = $endPeriod . '-31';
+        $fromDate = $startPeriod->format('Y-m-d');
+        $toDate = $endPeriod->format('Y-m-d');
 
         return $this->getResultsByDateRange($customerID, $fromDate, $toDate);
     }