+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Changelog
 This project changes will be shown here.
 
+## [unreleased]
+### Removed
+- Issue 241 - Remove Customer Review Meeting from the Accounts menu
+
+
 ## [v1.6.2] - 2018-05-21
 ### Fixed
 - Issue 234 - Permissions bug for sales order address page