+++ CHANGELOG.md
@@ -4,6 +4,7 @@ This project changes will be shown here.
 ## [unreleased]
 ### Added
 - Issue 246 - Add descriptions to Report pages
+- Issue 237 - ServiceRequestReport export tweak
 ### Changed
 - Issue 233 - Amend Create Renewal Sales Order email destination
 ### Removed